+++ linesman/parse.py
@@ -5,29 +5,29 @@ import gpxpy
 from .geometry import Vector, Line
 
 
-def lonlat_str(string):
+def latlon_str(string):
     """:return: Vector(lon, lat) point"""
     if ',' not in string:
-        raise ValueError("Format must be 'lon,lat' (missing ',')!")
-    lon, lat = string.split(',', maxsplit=1)
+        raise ValueError("Format must be 'lat,lon' (missing ',')!")
+    lat, lon = string.split(',', maxsplit=1)
     try:
         lon = float(lon)
-    except ValueError as e:
+    except ValueError:
         raise ValueError(f"lon '{lon}' is no valid floating point number.")
     try:
         lat = float(lat)
-    except ValueError as e:
+    except ValueError:
         raise ValueError(f"lat '{lat}' is no valid floating point number.")
     return Vector(lon, lat)
 
 
-def lonlat_pair_str(string):
+def latlon_pair_str(string):
     """:return: Line instance"""
     if ';' not in string:
         raise ValueError("Format for line must be 'start;end' (missing ';')!")
 
     start, end = string.split(';', maxsplit=1)
-    return Line(lonlat_str(start), lonlat_str(end))
+    return Line(latlon_str(start), latlon_str(end))
 
 
 def gpx_file(path):